api/requirements.txt
gunicorn==20.1.0
mysqlclient==2.1.0
python-dotenv==0.19.2
Django==4.0.1
djangorestframework==3.13.1
django-cors-headers==3.11.0
django-filter==2.4.0
djoser==2.1.0

# django-extensions==3.1.5
# django-storages==1.12.3
# django-anymail==8.5
# redis==3.5.3
# celery==5.2.3
# django-celery-results==2.1.0
# django-celery-beat==2.2.0
# Faker==13.0.0
# Pillow==9.0.1
# premailer==3.10.0